app.py
import streamlit as st
import pandas as pd
from fpdf import FPDF
from langchain_google_genai import ChatGoogleGenerativeAI
from langchain.prompts import PromptTemplate
from dotenv import load_dotenv
import os

load_dotenv()

# --- CONFIGURATION ---
st.set_page_config(page_title="AdPulse | Automated Insights", page_icon="📊", layout="centered")

# --- Gemini API KEY from ENV ---
api_key = os.getenv("GEMINI_API_KEY")

# --- CORE LOGIC (PANDAS) ---

# --- AI INSUGHT GENERATOR ---

# --- PDF GENERATION ---

# --- MAIN APPLICATION UI ---
def main():
    st.title("📊 AdPulse: Automated Insight Engine")
    st.markdown("""
    **Transform raw campaign data into executive decisions in seconds.**
    Upload your weekly CSV to generate a PDF report with AI-driven analysis.
    """)
    
    # 1. File Upload
    
    # 2. Process Data
    

if __name__ == "__main__":
    main()