requirements.txt
streamlit
pandas
langchain
langchain-google-genai
fpdf
python-dotenv